// app_aeroporto/src/Interfaces/types.ts
export interface FormValues {
  id: string;
  destiny: string;
  date: string;
  duration: string;
  price: string;
  attractions: string;
  photo: string;
}

export interface DataUser {
  id: string;
  destiny: string;
  date: string;
  duration: string;
  price: string;
  attractions: string;
  photo: string;
}